+++ src/State.h
@@ -22,15 +22,15 @@ typedef struct State {
   STATE_ID_T id;
   char name[STATE_NAME_LENGTH];
   struct EmissionTable* emission_table;
-  uint8_t num_emissions;
+  uint16_t num_emissions;
   Literal reference[STATE_MAX_REF_LEN];
-  uint8_t num_incoming;
+  uint16_t num_incoming;
   struct Transition incoming[STATE_MAX_NUM_INCOMING];
   STATE_CUSTOM_T custom;
 } State;
 
-bool State__init(struct State* self, char name[STATE_NAME_LENGTH], uint8_t num_emissions, Literal emission_reference[num_emissions], struct EmissionTable* emission_table);
-bool State__init_uniform(struct State* self, char name[STATE_NAME_LENGTH], uint8_t num_emissions, struct EmissionTable* emission_table);
+bool State__init(struct State* self, char name[STATE_NAME_LENGTH], uint16_t num_emissions, Literal emission_reference[num_emissions], struct EmissionTable* emission_table);
+bool State__init_uniform(struct State* self, char name[STATE_NAME_LENGTH], uint16_t num_emissions, struct EmissionTable* emission_table);
 bool State__init_silent(struct State* self, char name[STATE_NAME_LENGTH]);
 bool State__add_incoming(struct State* self, LOGODD_T logodd, struct State* target);
 bool State__str(struct State* self, char* buffer);
